# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GinAdmin 示例项目",
        "title": "GinAdmin Api",
        "contact": {
            "name": "gphper",
            "url": "https://github.com/kumataahh/eagle-eye"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:20011",
    "basePath": "/api",
    "paths": {
        "/example/index": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "示例"
                ],
                "summary": "展示用户信息",
                "operationId": "1",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "年龄",
                        "name": "age",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "性别",
                        "name": "sex",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.UserReq"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/api.DefaultResponse"
                        }
                    }
                }
            }
        },
        "/user//refresh": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "刷新jtoken",
                "operationId": "4",
                "parameters": [
                    {
                        "type": "string",
                        "name": "retoken",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\":1,\"msg\":\"success\",\"data\":{\"jtoken\":\"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJFeHAiOiIyMDIxLTEyLTI2VDE5OjI1OjI4Ljg0OTIzNzUrMDg6MDAiLCJOYW1lIjoiZ3BocGVyIiwiVWlkIjo0fQ==.ab81bb7134978afe976df55b45789aefd10f6c3edb969bae283c32c080083b89\"}}",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/api.DefaultResponse"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录",
                "operationId": "3",
                "parameters": [
                    {
                        "type": "string",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserLoginRes"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/api.DefaultResponse"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户注册",
                "operationId": "2",
                "parameters": [
                    {
                        "type": "string",
                        "name": "confirm_password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "nickname",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SuccessResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/api.DefaultResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.DefaultResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "code 为1表示正常 0表示业务请求错误",
                    "type": "integer"
                },
                "msg": {
                    "description": "错误提示信息",
                    "type": "string"
                }
            }
        },
        "api.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "code 为1表示正常 0表示业务请求错误",
                    "type": "integer"
                },
                "data": {
                    "description": "接口返回的业务数据",
                    "type": "object"
                },
                "msg": {
                    "description": "错误提示信息",
                    "type": "string"
                }
            }
        },
        "models.UserLoginRes": {
            "type": "object",
            "properties": {
                "jtoken": {
                    "description": "Jtoken 验证字符串",
                    "type": "string"
                },
                "retoken": {
                    "description": "retoken 刷新token",
                    "type": "string"
                }
            }
        },
        "models.UserReq": {
            "type": "object",
            "required": [
                "age",
                "sex",
                "username"
            ],
            "properties": {
                "age": {
                    "description": "年龄",
                    "type": "integer"
                },
                "sex": {
                    "description": "性别",
                    "type": "integer"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        }
    }
}